+++ modules/audioset/src/migrate.js
@@ -1,7 +1,7 @@
 import hsvToRgb from "hsv-rgb";
 
 const HSV = /^hsv\((.*)\)$/;
-export function rgb(color) {
+function rgb(color) {
   const parsed = HSV.exec(color)[1]
     .split(",")
     .map(i => parseInt(i, 10));
@@ -9,12 +9,22 @@ export function rgb(color) {
   return "rgb(" + c[0] + "," + c[1] + "," + c[2] + ")";
 }
 
+function resourceUrl(name, resource) {
+  const url =
+    resource && resource.length
+      ? resource[0].replace("{{filename}}", name)
+      : "";
+  return url;
+}
+
 export default function migrate(set) {
-  console.log("migrate set!", set);
+  console.log("migrate", set.id);
   Object.keys(set.clips).forEach(name => {
     const clip = set.clips[name];
     clip.id = name;
     clip.display.color = rgb(clip.display.color);
+    clip.display.cover = resourceUrl(name, set.loader.sources.covers);
+    console.log(clip.display);
   });
   return set;
 }